% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_depot}
\alias{add_depot}
\title{Add a new depot}
\usage{
add_depot(username, broker, external_id, ccy)
}
\arguments{
\item{username}{The owner's user name}

\item{broker}{The portfolio's broker / bank}

\item{external_id}{The portfolio's ID with the broker}

\item{ccy}{The depot's currency.}
}
\description{
Add a new depot
}
\examples{
\dontrun{
add_depot('me','my bank','portfolio 123132','EUR')
}
}
