% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_instrument}
\alias{add_instrument}
\title{Introduce a new security}
\usage{
add_instrument(isin, name, ccy)
}
\arguments{
\item{isin}{International Securities Idenficiation Number of the new object,
e.g. 'DE1234567890'}

\item{name}{A name, e.g. 'Acme Inc Pref. Shares'}

\item{ccy}{The company's / fund's reporting currency, e.g. 'EUR'}
}
\description{
Introduce a new security
}
\examples{
\dontrun{
add_instrument('DE1234567890','Acme Inc','EUR')
}
}
