% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_or_update_symbol}
\alias{add_or_update_symbol}
\title{Introduce a new ticker symbol}
\usage{
add_or_update_symbol(vendor, mic, isin, ccy, symbol)
}
\arguments{
\item{vendor}{Name of the vendor that's providing the symbol, e.g. 'IEX'}

\item{mic}{The symbol's market, e.g. 'XETR', 'NYSE'}

\item{isin}{International Securities Identification Number of the symbol, e.g.
'DE1234566780'}

\item{ccy}{The trading currency, e.g. 'EUR'}

\item{symbol}{The actual symbol, e.g. 'ALV.DE' (Alphavantage) or 'ALV-GY' (IEX')}
}
\value{

}
\description{
Introduce a new ticker symbol
}
\examples{
\dontrun{
add_or_update_symbol('IEX','XETR','DE1234567879','EUR','MYSYMB-GY')
}
}
