% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookkeeping.R, R/functions.R
\name{add_user}
\alias{add_user}
\title{Add a new user to the system}
\usage{
add_user(username, password)

add_user(username, password)
}
\arguments{
\item{username}{The new username}

\item{password}{The new password}
}
\description{
Add a new user to the system

Add a new user to the system
}
\examples{
\dontrun{
add_user('me','mypass')
}
\dontrun{
add_user('me','mypass')
}
}
