% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{authenticate}
\alias{authenticate}
\title{Credentials for user authentification}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{Your depot user}

\item{password}{Your depot password}
}
\value{

}
\description{
Credentials for user authentification
}
\examples{
\dontrun{
authenticate('me','mypass')
}
}
