% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{book_ticket}
\alias{book_ticket}
\title{Book a trade or casfhlow ticket against the backend}
\usage{
book_ticket(broker, external_id, valuedate, ...)
}
\arguments{
\item{broker}{The broker's name}

\item{external_id}{The depot's ID/name with the broker.}

\item{valuedate}{the value date of the ticket, e.g. trade date + 2.}

\item{...}{The actual ticket content. At the moment, we accept `cashflow` and
`trade` tickets. See examples.}
}
\value{
To be implemented.
}
\description{
Book a trade or casfhlow ticket against the backend
}
\examples{
\dontrun{
initDB('localhost',3306,'batch_user','batch_password')

# a trade with a single payment attached. Payment is optional:
book_ticket(
  broker        = 'bank inc',
  external_id = 'BI_adam_123',
  valuedate     = '2021-05-10',
  isin          = "DE1234567890",
  trade         = list(qty = 100,
                       prc = 999,
                       ccy = "EUR"),
  payment       = list(type   = "tax",
                       amount = 999,
                       ccy    = "EUR"))

# Trade with two payments
book_ticket(
  broker        = 'bank inc',
  external_id = 'BI_adam_123',
  valuedate     = '2021-05-10',
  isin          = "DE1234567890",
  trade         = list(qty = 100,
                       prc = 999,
                       ccy = "EUR"),
  payment       = list(
                     list(type   = "tax",
                          amount = 999,
                          ccy    = "EUR"),
                     list(type   = "fee",
                          amount = 100
                          ccy    = "EUR")))

 Dividend type cashflow
book_ticket(
  broker        = 'bank inc',
  external_id = 'BI_adam_123',
  valuedate     = '2021-05-10',
  isin          = "DE1234567890",
  cashflow      = list(type   = "dividend/distribution",
                       amount = 50,
                       ccy    = "EUR"))
}
}
