% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exec_wrapper}
\alias{exec_wrapper}
\title{SQL execute wrapper}
\usage{
exec_wrapper(sql_statement, authenticate = FALSE)
}
\arguments{
\item{sql_statement}{A prepared SQL statement}
}
\value{

}
\description{
SQL execute wrapper
}
\examples{
\dontrun{
exec_wrapper("CALL add_user('myname');")
}
}
