% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_corporate_actions}
\alias{extract_corporate_actions}
\title{Extract corporate actions from (un-)adjusted closing prices}
\usage{
extract_corporate_actions(prices)
}
\arguments{
\item{prices}{A tibble with `symbol`, `date`, unadjusted close `uClose`,
split adjusted close `close` and fully-adjusted `fClose`.}
}
\value{
A tibble with `symbol`, `date`,`close`, `split_factor` and `dividend`
}
\description{
Extract corporate actions from (un-)adjusted closing prices
}
\examples{
prices <- tibble::tibble(
    symbol = c("ALV-GY","ALV-GY","ALV-GY"),
      date = c("2021-05-04","2021-05-05","2021-05-06"),
     close = c(216.5000, 221.5000, 212.7000),
    aClose = c(216.5000, 221.5000, 212.7000),
    fClose = c(207.1167, 211.9000, 212.7000))
extract_corporate_actions(prices)
}
