% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flow_table}
\alias{flow_table}
\title{Cash flows to/from your depot(s) on isin/ccy level, assuming
initial position to be an investment at market prices (neg. sign)}
\usage{
flow_table(from, to)
}
\arguments{
\item{from}{Start date, inclusive.}

\item{to}{End date, inclusive.}
}
\value{
A tibble with depot and isin identifiers, cash flow size and ccy.
}
\description{
Cash flows to/from your depot(s) on isin/ccy level, assuming
initial position to be an investment at market prices (neg. sign)
}
\examples{
\dontrun{flow_table('2021-01-01','2021-05-31')}
}
