% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_connection}
\alias{get_connection}
\title{Provide a connection object}
\usage{
get_connection()
}
\value{
A function holding the `connection` and its `close` method.
}
\description{
Provide a connection object
}
