% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_market}
\alias{get_market}
\title{Obtain daily OHLC prices and volumes per isin/ccy.}
\usage{
get_market(isins, from, to, ...)
}
\arguments{
\item{isins}{Optional list of ISINs.}

\item{from}{Optional start date, defaults to '1999-12-31'}

\item{to}{Optional end date, defaults to today.}

\item{...}{Optional filter parameters passed to the backend}
}
\value{
A tibble with `isin`, `ccy`, `valuedate`, ohlc, fully adjusted close `fClose`, and volume.
}
\description{
Obtain daily OHLC prices and volumes per isin/ccy.
}
\examples{
\dontrun{
get_market()

get_market("DE0008404005")
get_market(,,,valuedate=="2021-05-05")
}
}
