% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_wrapper}
\alias{get_wrapper}
\title{SQL get wrapper}
\usage{
get_wrapper(sql_statement, authenticate = FALSE)
}
\arguments{
\item{sql_statement}{A prepared SQL statement}
}
\value{

}
\description{
SQL get wrapper
}
\examples{
\dontrun{
get_wrapper("SELECT * FROM v_last_symbol_data;")
}
}
