% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grant_depot_permission}
\alias{grant_depot_permission}
\title{Grant, or revoke, read/write permissions to your portfolios}
\usage{
grant_depot_permission(
  to_user,
  broker,
  external_id,
  permission_level = c("revoke", "read", "write")
)
}
\arguments{
\item{to_user}{The grantee}

\item{broker}{The depot's brokerage}

\item{external_id}{The depot's id with your broker}

\item{permission_level}{The targeted permission level, must be 'revoke', 'read' or 'write'}
}
\description{
Grant, or revoke, read/write permissions to your portfolios
}
\examples{
\dontrun{
grant_depot_permission('you','my broker','MB12311/2','read')
}
}
