% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initDB}
\alias{initDB}
\title{Initialize the MariaDB backend database}
\usage{
initDB(host, port, user, pass, dbname, ...)
}
\arguments{
\item{host}{Host address, e.g. 'localhost' or 'IP address'}

\item{port}{Backend port, e.g. 3306}

\item{user}{User name}

\item{pass}{Password}

\item{dbname}{Database name}

\item{...}{Other parameters to be passed to the backend connection}
}
\value{
Null
}
\description{
Initialize the MariaDB backend database
}
\examples{
\dontrun{
initDB('localhost',3306,'adam','the password')
}
}
