% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{latest_prices}
\alias{latest_prices}
\title{Download latest symbol data from IEX cloud}
\usage{
latest_prices()
}
\value{
A tibble with `vendor`, `symbol`, `date`, `OHLC`, and volume
}
\description{
Download latest symbol data from IEX cloud
}
\examples{
\dontrun{
latest_prices()
}
}
