% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{performance_table}
\alias{performance_table}
\title{Internal rate of return based on flow table}
\usage{
performance_table(flows)
}
\arguments{
\item{flows}{A tibble from `flow_table`.}
}
\value{

}
\description{
Internal rate of return based on flow table
}
\examples{
\dontrun{
flow_table('2021-01-01','2021-05-31')
performance_table(flows)}
}
