% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{position}
\alias{position}
\title{Get position table for active user}
\usage{
position(valuedate = Sys.Date())
}
\arguments{
\item{valuedate}{As of which date is the position to be returned?
Defaults to _today_.}
}
\value{
A tibble with depot properties (owner,broker, external_id, ccy),
instrument properties (isin,ccy), position properties (date, qty, vol, ccy)
}
\description{
Get position table for active user
}
\examples{
\dontrun{
authenticate('me','mypass')
position()
}
}
