% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{price_insert_batch}
\alias{price_insert_batch}
\title{Batch insert latest market data into the backend}
\usage{
price_insert_batch()
}
\description{
Batch insert latest market data into the backend
}
\examples{
\dontrun{
iexcloudr::apikey("my_api_key")
depotr::initDB('localhost',3399,'user','pass')
price_insert_batch()
}
}
