% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{store_corporate_actions}
\alias{store_corporate_actions}
\title{Store price adjustments from corporate actions}
\usage{
store_corporate_actions(actions)
}
\arguments{
\item{actions}{A tibble with `vendor`, `symbol`, `date`,`close`, `dividend` and `split_factor`.}
}
\description{
Store price adjustments from corporate actions
}
\examples{
\dontrun{
iexcloudr::prices("ALV-GY") \%>\%
iexcloudr:extract_corporate_actions() \%>\%
store_corporate_actions()}
}
