% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{store_prices}
\alias{store_prices}
\title{Load new market data to database, based on a data tibble.}
\usage{
store_prices(prices)
}
\arguments{
\item{prices}{A tibble with `vendor`, `symbol`, `date`, `OHLC`, and volume}
}
\description{
Load new market data to database, based on a data tibble.
}
\examples{
\dontrun{
store_prices(pricetbl)}
}
