% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trades}
\alias{trades}
\title{Get list of trades for active user}
\usage{
trades()
}
\value{
A tibble with depot properties (owner,broker, external_id, ccy),
instrument properties (isin,ccy), trade properties (date, qty, prc, vol, ccy)
}
\description{
Get list of trades for active user
}
\examples{
\dontrun{
authenticate('me','mypass')
trades()
}
}
