% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookkeeping.R
\name{cashflows}
\alias{cashflows}
\title{Get all portfolio cashflows for a time period}
\usage{
cashflows(from_date = "1990-01-01", to_date = Sys.Date())
}
\arguments{
\item{from_date}{Start date, inclusive. Defaults to first of the year}

\item{to_date}{End date, inclusive. Defaults to 'today'.}
}
\value{
A tibble with depot and instrument identifiers,
cash flow properties and amounts.
}
\description{
Get all portfolio cashflows for a time period
}
